+++ src/main/kotlin/net/ndrei/teslapoweredthingies/machines/powdermaker/PowderMakerRecipes.kt
@@ -1,12 +1,10 @@
 package net.ndrei.teslapoweredthingies.machines.powdermaker
 
 import net.minecraft.init.Blocks
+import net.minecraft.init.Items
 import net.minecraft.item.ItemStack
 import net.minecraftforge.oredict.OreDictionary
-import net.ndrei.teslapoweredthingies.common.IRecipeOutput
-import net.ndrei.teslapoweredthingies.common.OreOutput
-import net.ndrei.teslapoweredthingies.common.SecondaryOreOutput
-import net.ndrei.teslapoweredthingies.common.SecondaryOutput
+import net.ndrei.teslapoweredthingies.common.*
 
 /**
  * Created by CF on 2017-07-05.
@@ -35,7 +33,10 @@ object PowderMakerRecipes {
 
     fun registerDefaultOreRecipe(oreKey: String, input: ItemStack, isOre: Boolean) {
         val recipe = this.findDefaultOreRecipe(oreKey, input, isOre) ?: return
-        this.registerRecipe(recipe)
+
+        if (recipe.ensureValidOutputs().isNotEmpty()) {
+            this.registerRecipe(recipe)
+        }
     }
 
     fun registerDefaultOreRecipe(oreKey: String) {
@@ -53,37 +54,45 @@ object PowderMakerRecipes {
     }
 
     private fun addDefaultOreRecipe(key: String, primaryOutput: String, primaryOutputCount: Int, vararg output: IRecipeOutput) {
-        this.defaultOreRecipes[key] = { it, ore ->
+        this.addDefaultOreRecipe(key, { it, ore ->
             PowderMakerRecipe(it,
                     OreOutput(primaryOutput, primaryOutputCount),
                     *if (ore) arrayOf(SecondaryOutput(0.15f, Blocks.COBBLESTONE), *output)
                     else output
             )
-        }
+        })
+    }
+
+    private fun addDefaultOreRecipe(key: String, builder: (input: ItemStack, isOre: Boolean) -> IPowderMakerRecipe) {
+        this.defaultOreRecipes[key] = builder
     }
 
     init {
         this.addDefaultOreRecipe("iron",
-                SecondaryOreOutput(0.05f, "dustTin", 1),
-                SecondaryOreOutput(0.1f, "dustNickel", 1))
+            SecondaryOreOutput(0.05f, "dustTin", 1),
+            SecondaryOreOutput(0.1f, "dustNickel", 1))
         this.addDefaultOreRecipe("gold")
         this.addDefaultOreRecipe("copper",
-                SecondaryOreOutput(0.125f, "dustGold", 1))
+            SecondaryOreOutput(0.125f, "dustGold", 1))
         this.addDefaultOreRecipe("tin")
         this.addDefaultOreRecipe("silver",
-                SecondaryOreOutput(0.1f, "dustLead", 1))
+            SecondaryOreOutput(0.1f, "dustLead", 1))
         this.addDefaultOreRecipe("lead",
-                SecondaryOreOutput(0.1f, "dustSilver", 1))
+            SecondaryOreOutput(0.1f, "dustSilver", 1))
         this.addDefaultOreRecipe("aluminum")
         this.addDefaultOreRecipe("nickel",
-                SecondaryOreOutput(0.1f, "dustPlatinum", 1))
+            SecondaryOreOutput(0.1f, "dustPlatinum", 1))
         this.addDefaultOreRecipe("platinum")
         this.addDefaultOreRecipe("iridium")
 
-        this.addDefaultOreRecipe("coal", "coal", 6)
-        this.addDefaultOreRecipe("diamond", "gemDiamond", 4)
-        this.addDefaultOreRecipe("emerald", "gemEmerald", 4)
+        this.addDefaultOreRecipe("coal", { it, ore ->
+            PowderMakerRecipe(it, *mutableListOf<IRecipeOutput>(Output(ItemStack(Items.COAL))).also { when (ore) {
+                true -> it.add(SecondaryOutput(0.15f, Blocks.COBBLESTONE))
+            }}.toTypedArray())
+        })
+        this.addDefaultOreRecipe("diamond", "dustDiamond", 3)
+        this.addDefaultOreRecipe("emerald", "dustEmerald", 3)
         this.addDefaultOreRecipe("redstone", "dustRedstone", 6)
-        this.addDefaultOreRecipe("lapis", "gemLapis", 6)
+        this.addDefaultOreRecipe("lapis", "gemLapis", 5)
     }
 }