+++ changelog.md
@@ -11,3 +11,4 @@
     - check "-base.json" files for examples of the recipe structure
 - improved liquid xp storage machine 
 - added 'player xp addon' - you put it into a 'liquid xp storage' and you can convert stored liquid xp to/from player xp
+- fixed some of the default recipes for 'powder maker'